perimeter

+++ src/area.rs
@@ -3,5 +3,5 @@ use rocket::serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct Area {
-    pub area: f32,
+    pub area: f64,
 }
